4.21*: bloglist expansion, step9
- Added code to ensure only the user who created a blog post can delete it

+++ part4/bloglist/controllers/blogs.js
@@ -49,6 +49,18 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET);
+    
+    if (!request.token || !decodedToken.id) {
+        return response.status(401).json({ error: 'token missing or invalid' });
+    }
+    
+    const blogToRemove = await Blog.findById(request.params.id);
+    
+    if(decodedToken.id !== blogToRemove.user.toString()) {
+        return response.status(403).json({ error: 'only the user that created the blog post can delete it' });  
+    }
+
     await Blog.findByIdAndRemove(request.params.id);
     response.status(204).end();
 })
